+++ src/entities/settings.py
@@ -8,6 +8,7 @@ class SettingsDB(DatabaseFile):
         DatabaseFile.__init__(self, converter, "settings.db")
         self.entities = [Setting(self, "core.moduleConfiguration", {"permission_viewer":True}),
                          Setting(self, "dnd5e.systemMigrationVersion", foundry.DEFAULT_SYSTEM_VERSION),
+                         Setting(self, "dnd5e.rulesVersion", "legacy"),
                          Setting(self, "permission_viewer.migrated", "1"),
                          Setting(self, "core.permissions", {})]